requirements.txt
pandas>=2.0
numpy>=1.24
python-dateutil>=2.9
tqdm>=4.66
matplotlib>=3.8
plotly>=5.24
scikit-learn>=1.4
vaderSentiment>=3.3.2
Mastodon.py>=1.8.1
praw>=7.7.1
streamlit>=1.38
langdetect>=1.0.9
transformers>=4.44.2
sentencepiece>=0.2.0
torch>=2.1.0
sacremoses>=0.0.53